requirements.txt
Django==2.2.5
django-rest-swagger==2.2.0
djangorestframework==3.10.2
drf-yasg==1.16.1
PyMySQL==0.9.3
absl-py==0.8.1
astor==0.8.0
boto==2.49.0
boto3==1.9.246
botocore==1.12.246
certifi==2019.9.11
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
cycler==0.10.0
DBUtils==1.3
decorator==4.4.1
docutils==0.15.2
drf-yasg==1.16.1
elasticsearch==7.0.4
gast==0.2.2
gensim==3.8.1
google-pasta==0.1.7
grpcio==1.24.3
h5py==2.10.0
idna==2.8
inflection==0.3.1
itypes==1.1.0
jieba==0.39
Jinja2==2.10.1
jmespath==0.9.4
joblib==0.14.0
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
Markdown==3.1.1
MarkupSafe==1.1.1
matplotlib==3.1.1
networkx==2.4
numpy==1.17.2
openapi-codec==1.3.2
opt-einsum==3.1.0
pandas==0.25.2
pip2pi==0.8.1
protobuf==3.10.0
pyparsing==2.4.2
python-dateutil==2.8.0
pytz==2019.2
PyYAML==5.1.2
redis==3.3.8
requests==2.22.0
ruamel.yaml==0.16.5
ruamel.yaml.clib==0.1.2
s3transfer==0.2.1
scikit-learn==0.21.3
scipy==1.3.1
simplejson==3.16.0
six==1.12.0
sklearn==0.0
smart-open==1.8.4
sqlparse==0.3.0
tensorboard==2.0.0
tensorflow==2.0.0
tensorflow-estimator==2.0.1
termcolor==1.1.0
uritemplate==3.0.0
urllib3==1.25.3
Werkzeug==0.16.0
wrapt==1.11.2